+++ app/AgedBrieItem.ts
@@ -1,0 +1,16 @@
+import { BaseItem } from "./BaseItem";
+
+
+export class AgedBrieItem extends BaseItem {
+  updateQuality() {
+    if (!this.isMaxQuality()) {
+      this.upgradeItemUnit();
+    }
+
+    this.decreaseSellInUnit();
+
+    if (this.hasExpired() && !this.isMaxQuality()) {
+      this.upgradeItemUnit();
+    }
+  }
+}
