+++ app/gilded-rose.ts
@@ -1,131 +1,4 @@
-export class Item {
-  name: string;
-  sellIn: number;
-  quality: number;
-
-  constructor(name, sellIn, quality) {
-    this.name = name;
-    this.sellIn = sellIn;
-    this.quality = quality;
-  }
-}
-
-export class BaseItem extends Item {
-  protected hasExpired() {
-    return this.sellIn < 0;
-  }
-
-  protected decreaseSellInUnit() {
-    this.sellIn = this.sellIn - 1;
-  }
-
-  protected isMaxQuality() {
-    return this.quality === GildedRose.MAX_QUALITY;
-  }
-
-  protected modifyItemQuality(amount: number) {
-    this.quality = this.quality + amount;
-  }
-
-  protected upgradeItemUnit() {
-    this.modifyItemQuality(1);
-  }
-
-  protected downgradeItemUnit() {
-    this.modifyItemQuality(-1);
-  }
-
-  updateQuality() {
-    if (this.quality > 0) {
-      this.downgradeItemUnit();
-    }
-
-    this.decreaseSellInUnit();
-
-    if (this.hasExpired() && this.quality > 0) {
-      this.downgradeItemUnit();
-    }
-  }
-}
-
-export class AgedBrieItem extends BaseItem {
-  updateQuality() {
-    if (!this.isMaxQuality()) {
-      this.upgradeItemUnit();
-    }
-
-    this.decreaseSellInUnit();
-
-    if (this.hasExpired() && !this.isMaxQuality()) {
-      this.upgradeItemUnit();
-    }
-  }
-}
-
-export class SulfurasItem extends BaseItem {
-  private static readonly DEFAULT_QUALITY = 80;
-
-  constructor (name: string) {
-    super(name, 0, SulfurasItem.DEFAULT_QUALITY);
-  }
-
-  updateQuality() {
-  }
-}
-
-export class BackstagePassItem extends BaseItem {
-  updateQuality() {
-    if (!this.isMaxQuality()) {
-      this.upgradeItemUnit();
-
-      if (this.sellIn <= 10 && !this.isMaxQuality()) {
-        this.upgradeItemUnit();
-      }
-
-      if (this.sellIn <= 5 && !this.isMaxQuality()) {
-        this.upgradeItemUnit();
-      }
-    }
-
-    this.decreaseSellInUnit();
-
-    if (this.hasExpired()) {
-      this.modifyItemQuality(-this.quality);
-    }
-  }
-}
-
-export class ConjuredItem extends BaseItem {
-  updateQuality() {
-      if (this.quality > 0) {
-        this.modifyItemQuality(-2);
-          if (this.quality < 0) this.quality = 0;
-      }
-
-      this.decreaseSellInUnit()
-
-      if (this.hasExpired() && this.quality > 0) {
-        this.modifyItemQuality(-2);
-          if (this.hasExpired()) this.quality = 0;
-      }
-  }
-}
-
-export class ItemFactory {
-  static createItem(name: string, sellIn: number, quality: number): BaseItem {
-      if (name === 'Aged Brie') {
-          return new AgedBrieItem(name, sellIn, quality);
-      } else if (name === 'Sulfuras, Hand of Ragnaros') {
-          return new SulfurasItem(name);
-      } else if (name === 'Backstage passes to a TAFKAL80ETC concert') {
-          return new BackstagePassItem(name, sellIn, quality);
-      } else if (name.startsWith('Conjured')) {
-          return new ConjuredItem(name, sellIn, quality);
-      } else {
-          return new BaseItem(name, sellIn, quality);
-      }
-  }
-}
+import { BaseItem } from "./BaseItem";
 
 export class GildedRose {
   public static readonly MAX_QUALITY: number = 50;
